+++ osbuild/formats/__init__.py
@@ -1,0 +1,3 @@
+"""
+Concrete representation of manifest descriptions
+"""
